HH-218702 move classes from nab-hibernate to nab-data-source module
