+++ nab-data-source/src/main/java/ru/hh/nab/datasource/aspect/ExecuteOnDataSourceWrappedException.java
@@ -1,8 +1,8 @@
-package ru.hh.nab.hibernate.transaction;
+package ru.hh.nab.datasource.aspect;
 
 import static java.util.Objects.requireNonNull;
 
-final class ExecuteOnDataSourceWrappedException extends RuntimeException {
+public final class ExecuteOnDataSourceWrappedException extends RuntimeException {
   public ExecuteOnDataSourceWrappedException(Throwable cause) {
     super("Checked exception from @ExecuteOnDataSource", requireNonNull(cause));
   }
