+++ nab-data-source/src/main/java/ru/hh/nab/datasource/routing/DataSourceContext.java
@@ -1,8 +1,7 @@
-package ru.hh.nab.hibernate.transaction;
+package ru.hh.nab.datasource.routing;
 
 import java.util.function.Supplier;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import ru.hh.nab.datasource.DataSourceType;
 
 public class DataSourceContext {
