+++ nab-data-source/src/main/java/ru/hh/nab/datasource/routing/DataSourceRoutingFilter.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.routing;
+package ru.hh.nab.datasource.routing;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
@@ -11,9 +11,7 @@ import java.io.IOException;
 import org.glassfish.jersey.uri.UriComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import ru.hh.nab.datasource.DataSourceType;
-import ru.hh.nab.hibernate.transaction.DataSourceContext;
 
 public class DataSourceRoutingFilter implements Filter {
   private static final Logger LOG = LoggerFactory.getLogger(DataSourceRoutingFilter.class);
