+++ nab-data-source/src/main/java/ru/hh/nab/datasource/routing/RoutingDataSourceFactory.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.datasource;
+package ru.hh.nab.datasource.routing;
 
 import jakarta.inject.Named;
 import javax.sql.DataSource;
