+++ nab-data-source/src/main/java/ru/hh/nab/datasource/validation/ExecuteOnDataSourceBeanPostProcessor.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.transaction;
+package ru.hh.nab.datasource.validation;
 
 import java.lang.reflect.Method;
 import java.util.Collection;
@@ -11,9 +11,10 @@ import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.core.annotation.AnnotationConfigurationException;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import ru.hh.nab.datasource.DataSourcePropertiesStorage;
 import ru.hh.nab.datasource.DataSourceType;
+import ru.hh.nab.datasource.annotation.ExecuteOnDataSource;
+import ru.hh.nab.datasource.routing.DataSourceContextUnsafe;
 
 public class ExecuteOnDataSourceBeanPostProcessor implements BeanPostProcessor {
