+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceTransactionCallback.java
@@ -7,6 +7,9 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.springframework.lang.NonNull;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.TransactionCallback;
+import ru.hh.nab.datasource.annotation.DataSourceCacheMode;
+import ru.hh.nab.datasource.annotation.ExecuteOnDataSource;
+import ru.hh.nab.datasource.aspect.ExecuteOnDataSourceWrappedException;
 
 public class ExecuteOnDataSourceTransactionCallback implements TransactionCallback<Object> {
