+++ nab-tests/src/test/java/ru/hh/nab/datasource/DataSourceSwitchingTest.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.jpa;
+package ru.hh.nab.datasource;
 
 import com.codahale.metrics.health.HealthCheck;
 import com.zaxxer.hikari.HikariConfig;
@@ -34,22 +34,21 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.transaction.annotation.Transactional;
 import ru.hh.nab.common.properties.FileSettings;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
-import ru.hh.nab.datasource.DataSourceContextUnsafe;
-import ru.hh.nab.datasource.DataSourceFactory;
-import ru.hh.nab.datasource.DataSourcePropertiesStorage;
 import static ru.hh.nab.datasource.DataSourceSettings.DATASOURCE_NAME_FORMAT;
 import static ru.hh.nab.datasource.DataSourceSettings.HEALTHCHECK_ENABLED;
 import static ru.hh.nab.datasource.DataSourceSettings.HEALTHCHECK_SETTINGS_PREFIX;
 import static ru.hh.nab.datasource.DataSourceSettings.ROUTING_SECONDARY_DATASOURCE;
-import ru.hh.nab.datasource.DataSourceType;
-import ru.hh.nab.datasource.DatabaseSwitcher;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSource;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.healthcheck.UnhealthyDataSourceException;
 import ru.hh.nab.datasource.monitoring.NabMetricsTrackerFactoryProvider;
-import ru.hh.nab.hibernate.datasource.RoutingDataSource;
-import ru.hh.nab.hibernate.datasource.RoutingDataSourceFactory;
-import static ru.hh.nab.hibernate.transaction.DataSourceContext.onDataSource;
+import static ru.hh.nab.datasource.routing.DataSourceContext.onDataSource;
+import ru.hh.nab.datasource.routing.DataSourceContextUnsafe;
+import ru.hh.nab.datasource.routing.DatabaseSwitcher;
+import ru.hh.nab.datasource.routing.RoutingDataSource;
+import ru.hh.nab.datasource.routing.RoutingDataSourceFactory;
+import ru.hh.nab.jpa.JpaTestConfig;
+import ru.hh.nab.jpa.MappingConfig;
 import ru.hh.nab.jpa.model.TestEntity;
 import ru.hh.nab.metrics.StatsDSender;
 import ru.hh.nab.testbase.jpa.JpaTestBase;
